refactor(py): rename vertex_ai package module to genkit_vertexai for consistency
