+++ py/packages/genkit-vertex-ai/src/genkit_vertexai/constants.py
