+++ py/packages/genkit-vertex-ai/src/genkit_vertexai/model_garden/anthropic.py
