+++ py/packages/genkit-vertex-ai/tests/model_garden/client_test.py
@@ -19,13 +19,13 @@
 from unittest.mock import MagicMock, patch
 
 import pytest
-from genkit_vertex_ai.model_garden.client import OpenAIClient
+from genkit_vertexai.model_garden.client import OpenAIClient
 
 
 @pytest.mark.asyncio
 @patch('google.auth.default')
 @patch('google.auth.transport.requests.Request')
-@patch('genkit_vertex_ai.model_garden.client._AsyncOpenAI')
+@patch('genkit_vertexai.model_garden.client._AsyncOpenAI')
 async def test_client_initialization_with_explicit_project_id(
     mock_openai_cls: MagicMock, mock_request_cls: MagicMock, mock_default_auth: MagicMock
 ) -> None:
@@ -51,7 +51,7 @@ async def test_client_initialization_with_explicit_project_id(
 @pytest.mark.asyncio
 @patch('google.auth.default')
 @patch('google.auth.transport.requests.Request')
-@patch('genkit_vertex_ai.model_garden.client._AsyncOpenAI')
+@patch('genkit_vertexai.model_garden.client._AsyncOpenAI')
 async def test_client_initialization_without_explicit_project_id(
     mock_openai_cls: MagicMock, mock_request_cls: MagicMock, mock_default_auth: MagicMock
 ) -> None:
