+++ py/tests/smoke/package_test.py
@@ -19,7 +19,7 @@
 from genkit_google_cloud import package_name as google_cloud_package_name
 from genkit_googleai import package_name as google_genai_package_name
 from genkit_ollama import package_name as ollama_package_name
-from genkit_vertex_ai import package_name as vertex_ai_package_name
+from genkit_vertexai import package_name as vertex_ai_package_name
 
 
 def test_package_names() -> None:
@@ -31,4 +31,4 @@ def test_package_names() -> None:
     assert google_cloud_package_name() == 'genkit_google_cloud'
     assert google_genai_package_name() == 'genkit_googleai'
     assert ollama_package_name() == 'genkit_ollama'
-    assert vertex_ai_package_name() == 'genkit_vertex_ai'
+    assert vertex_ai_package_name() == 'genkit_vertexai'
